// src/app/shared/models/customers/customer.model.ts
export interface Customer {
  id?: string;
  name: string;
  birthDate: string;
  identityExpedition: string;
  cpfCnpj: string;
  mothersName: string;
  fathersName: string;
  civilState: string;
  partnersName: string;
  partnersCPF: string;
  cellphone: string;
  phoneNumber: string;
  cep: string;
  address: string;
  number: string;
  district: string;
  city: string;
  state: string;
  complement: string;
  referrals: string;
  email: string;
  observation: string;
}